requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
flask-cors==4.0.0
openai>=1.31.0
numpy==2.2.5
tqdm==4.67.1
pdfplumber==0.8.0
gunicorn==20.1.0
PyPDF2>=3.0.0
tiktoken
readability-lxml
beautifulsoup4
requests
fuzzywuzzy
python-Levenshtein









